+++ PhysIO/code/plot/tapas_physio_plot_raw_physdata_diagnostics.m
@@ -24,7 +24,7 @@ isVerbose = verbose.level > 0;
 if isVerbose
     fh = tapas_physio_get_default_fig_params();
     verbose.fig_handles(end+1) = fh;
-    set(fh, 'Name','Diagnostics for raw physiological time series');
+    set(fh, 'Name','Preproc: Diagnostics for raw physiological time series');
     ah = subplot(2,1,1);
     
     if hasCardiacData