+++ sem/examples/tapas_sem_example_hier_estimate.m
@@ -1,11 +1,18 @@
-function [posterior] = tapas_sem_hier_example_inversion(model, param)
-%% Test 
+function [posterior, summary] = tapas_sem_example_hier_estimate(model, param)
+%% Example for inversion with a linear model for the prior. 
 %
-% fp -- Pointer to a file for the test output, defaults to 1
-%
+% Input
+%       model       -- String. Either seria or prosa
+%       param       -- String. Parametric distribution.
+% Output
+%       posterior   -- Structure. Contains the posterior estimates.
+%       summary     -- Table. Contains a table with a summary of the 
+%                      posterior.
+
 % aponteeduardo@gmail.com
-% copyright (C) 2015
+% copyright (C) 2018
 %
+
 
 n = 0;
 
@@ -37,12 +44,14 @@ case 'seria'
             ptheta.llh = @c_seria_multi_later;
         case 'wald'
             ptheta.llh = @c_seria_multi_wald;
+        otherwise
+            error('parametric function not defined')
     end
 
     ptheta.jm = [...
         eye(19)
         zeros(3, 8) eye(3) zeros(3, 8)];
-   ptheta.p0(11) = tapas_logit([0.005], 1);
+    ptheta.p0(11) = tapas_logit([0.005], 1);
 case 'prosa'
     ptheta = tapas_sem_prosa_ptheta(); % Choose at convinience.
     switch param
@@ -58,6 +67,8 @@ case 'prosa'
         ptheta.llh = @c_prosa_multi_later;
     case 'wald'
         ptheta.llh = @c_prosa_multi_wald;
+    otherwise
+        error('parametric function not defined')
     end
 
     ptheta.jm = [...
@@ -68,7 +79,7 @@ end
 
 pars = struct();
 
-pars.T = ones(4, 1) * linspace(0.1, 1, 8).^5;
+pars.T = ones(4, 1) * linspace(0.1, 1, 1).^5;
 pars.nburnin = 4000;
 pars.niter = 4000;
 pars.ndiag = 500;
@@ -78,11 +89,11 @@ pars.verbose = 1;
 display(ptheta);
 inference = struct();
 inference.kernel_scale = 0.1 * 0.1;
-tic
+
 posterior = tapas_sem_hier_estimate(data, ptheta, inference, pars);
-toc
 
 display(posterior);
+summary = tapas_sem_display_posterior(posterior);
 
 end